python: move OutputFormat from flux-jobs.py over to utils.py
OutputFormat provides nice, mostly generalized functionality for
generating legible tabular data with headers. Originally used for
formatting the output for flux-jobs.py, this functionality would also be
useful for upcoming flux-sched commands (i.e., flux-tree-helper.py).
Move the class so that it is publically importable.
